@@ -10,30 +10,56 @@ export default function Login({ onAuth }) {
   const [error, setError] = useState(null);
   const navigate = useNavigate();
 
-  const submit = async (e) => {
-    e.preventDefault();
-    try {
-      const payload = { username: username.trim(), password };
-      const res = await api.post('/auth/login', payload);
-      const { token, role } = res.data;
-      localStorage.setItem('token', token);
-      localStorage.setItem('role', role);
-      localStorage.setItem('username', username);
-      if (onAuth) onAuth({ username, role });
-      // Navigate based on role
-      if (role === 'ROLE_ADMIN') {
-        navigate('/admin/dashboard');
-      } else if (role === 'ROLE_COLLECTOR') {
-        navigate('/collector/dashboard');
-      } else {
-        navigate('/user/dashboard');
-      }
-    } catch (err) {
-      // show server response in console for debugging
-      console.debug('Login error response:', err?.response);
-      setError(err?.response?.data?.error || 'Login failed');
+const submit = async (e) => {
+  e.preventDefault();
+  setError(null);
+  
+  try {
+    const payload = { 
+      username: username.trim(), 
+      password: password.trim() 
+    };
+    
+    console.log('Attempting login with:', payload);
+    
+    const res = await api.post('/auth/login', payload);
+    const { token, role } = res.data;
+    
+    console.log('Login successful:', res.data);
+    
+    localStorage.setItem('token', token);
+    localStorage.setItem('role', role);
+    localStorage.setItem('username', username);
+    
+    if (onAuth) onAuth({ username, role });
+    
+    // Navigate based on role
+    if (role === 'ROLE_ADMIN') {
+      navigate('/admin/dashboard');
+    } else if (role === 'ROLE_COLLECTOR') {
+      navigate('/collector/dashboard');
+    } else {
+      navigate('/user/dashboard');
     }
-  };
+  } catch (err) {
+    console.error('Login error:', err);
+    
+    if (err.response) {
+      // Server responded with error status
+      console.error('Response data:', err.response.data);
+      console.error('Response status:', err.response.status);
+      setError(err.response.data?.error || `Error: ${err.response.status}`);
+    } else if (err.request) {
+      // Request made but no response received
+      console.error('No response received:', err.request);
+      setError('Cannot connect to server. Please try again.');
+    } else {
+      // Other errors
+      console.error('Error:', err.message);
+      setError('An unexpected error occurred');
+    }
+  }
+};
 
   return (
     <div className="auth-container">
