@@ -1,149 +1,73 @@
-// package com.csse.smartwaste.config;
-
-// import java.util.List;
-
-// import com.csse.smartwaste.security.JwtFilter;
-// import org.springframework.context.annotation.Bean;
-// import org.springframework.context.annotation.Configuration;
-// import jakarta.servlet.http.HttpServletResponse;
-// import org.springframework.security.config.http.SessionCreationPolicy;
-// import org.springframework.http.HttpMethod;
-// import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-// import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-// import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-// import org.springframework.security.crypto.password.PasswordEncoder;
-// import org.springframework.security.web.SecurityFilterChain;
-// import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
-// import org.springframework.web.cors.CorsConfiguration;
-// import org.springframework.web.cors.CorsConfigurationSource;
-// import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
-
-// @Configuration
-// @EnableWebSecurity
-// public class SecurityConfig {
-
-//   private final JwtFilter jwtFilter;
-
-//   public SecurityConfig(JwtFilter jwtFilter) {
-//     this.jwtFilter = jwtFilter;
-//   }
-
-//   @Bean
-//   public PasswordEncoder passwordEncoder() {
-//     return new BCryptPasswordEncoder();
-//   }
-
-//   @Bean
-//   public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
-//     // Configure CORS and security rules:
-//     http
-//       .csrf(csrf -> csrf.disable())
-//       .cors(cors -> cors.configurationSource(corsConfigurationSource()))
-//       .sessionManagement(sm -> sm.sessionCreationPolicy(SessionCreationPolicy.STATELESS))
-//       .authorizeHttpRequests(auth -> auth
-//         // Allow preflight requests
-//         .requestMatchers(HttpMethod.OPTIONS, "/**").permitAll()
-//         // Allow unauthenticated access to auth endpoints and health check
-//         .requestMatchers("/api/auth/**", "/api/ping").permitAll()
-//         // All other requests must be authenticated
-//         .anyRequest().authenticated()
-//       )
-//       // Exception handling for unauthorized attempts
-//       .exceptionHandling(ex -> ex.authenticationEntryPoint((req, res, authEx) -> {
-//         res.setStatus(HttpServletResponse.SC_UNAUTHORIZED);
-//         res.setContentType("application/json");
-//         res.getWriter().write("{\"error\":\"Unauthorized\"}");
-//       }));
-
-//     // Add JWT filter before username/password filter
-//     http.addFilterBefore(jwtFilter, UsernamePasswordAuthenticationFilter.class);
-
-//     return http.build();
-//   }
-
-//   @Bean
-//   public CorsConfigurationSource corsConfigurationSource() {
-//     CorsConfiguration configuration = new CorsConfiguration();
-//     configuration.setAllowedOrigins(List.of("http://localhost:3000", "http://127.0.0.1:3000"));
-//     configuration.setAllowedMethods(List.of("GET", "POST", "PUT", "DELETE", "OPTIONS"));
-//     configuration.setAllowedHeaders(List.of("*"));
-//     configuration.setAllowCredentials(true);
-//     UrlBasedCorsConfigurationSource source = new UrlBasedCorsConfigurationSource();
-//     source.registerCorsConfiguration("/**", configuration);
-//     return source;
-//   }
-// }
+// src/main/java/com/csse/smartwaste/config/SecurityConfig.java
 package com.csse.smartwaste.config;
 
+import java.util.List;
+
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.http.HttpMethod;
+import org.springframework.security.authentication.AuthenticationManager;
+import org.springframework.security.config.Customizer;
+import org.springframework.security.config.annotation.authentication.configuration.AuthenticationConfiguration;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.http.SessionCreationPolicy;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.security.web.SecurityFilterChain;
+import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
 import org.springframework.web.cors.CorsConfiguration;
-import org.springframework.web.cors.CorsConfigurationSource;
-import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
 
-import java.util.Arrays;
+import com.csse.smartwaste.security.JwtFilter;
+import com.csse.smartwaste.security.JwtUtil;
 
 @Configuration
-@EnableWebSecurity
 public class SecurityConfig {
 
+    @Value("${app.jwt.secret}")
+    private String jwtSecret;
+    @Value("${app.jwt.expiration-ms}")
+    private long jwtExpirationMs;
+
     @Bean
-    public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
+    public JwtUtil jwtUtil() {
+        return new JwtUtil(jwtSecret, jwtExpirationMs);
+    }
+
+    @Bean
+    public JwtFilter jwtFilter(JwtUtil jwtUtil) {
+        return new JwtFilter(jwtUtil);
+    }
+
+    @Bean
+    public SecurityFilterChain filterChain(HttpSecurity http, JwtFilter jwtFilter) throws Exception {
         http
-            .cors(cors -> cors.configurationSource(corsConfigurationSource()))
             .csrf(csrf -> csrf.disable())
-            .authorizeHttpRequests(authz -> authz
+            .cors(cors -> cors.configurationSource(request -> {
+                var c = new CorsConfiguration();
+                c.setAllowedOrigins(List.of("http://localhost:3000","http://127.0.0.1:3000"));
+                c.setAllowedMethods(List.of("GET","POST","PUT","DELETE","PATCH","OPTIONS"));
+                c.setAllowedHeaders(List.of("Authorization","Content-Type","Accept"));
+                c.setAllowCredentials(true);
+                c.setMaxAge(3600L);
+                return c;
+            }))
+            .authorizeHttpRequests(auth -> auth
                 .requestMatchers("/api/auth/**").permitAll()
-                .requestMatchers("/api/debug/**").permitAll()
-                .requestMatchers("/error").permitAll()
+                .requestMatchers(HttpMethod.OPTIONS, "/**").permitAll()
                 .anyRequest().authenticated()
             )
-            .sessionManagement(session -> session
-                .sessionCreationPolicy(SessionCreationPolicy.STATELESS)
-            );
-        
+            .httpBasic(Customizer.withDefaults());
+
+        // Register JWT filter
+        http.addFilterBefore(jwtFilter, UsernamePasswordAuthenticationFilter.class);
         return http.build();
     }
-    
+
     @Bean
-    public CorsConfigurationSource corsConfigurationSource() {
-        CorsConfiguration configuration = new CorsConfiguration();
-        configuration.setAllowedOrigins(Arrays.asList(
-            "http://localhost:3000", 
-            "http://127.0.0.1:3000",
-            "http://localhost:3001",
-            "http://127.0.0.1:3001"
-        ));
-        configuration.setAllowedMethods(Arrays.asList("GET", "POST", "PUT", "DELETE", "OPTIONS", "PATCH"));
-        configuration.setAllowedHeaders(Arrays.asList(
-            "Authorization",
-            "Content-Type",
-            "X-Requested-With",
-            "Accept",
-            "Origin",
-            "Access-Control-Request-Method",
-            "Access-Control-Request-Headers"
-        ));
-        configuration.setExposedHeaders(Arrays.asList(
-            "Authorization",
-            "Content-Type"
-        ));
-        configuration.setAllowCredentials(true);
-        configuration.setMaxAge(3600L);
-        
-        UrlBasedCorsConfigurationSource source = new UrlBasedCorsConfigurationSource();
-        source.registerCorsConfiguration("/**", configuration);
-        return source;
-    }
-    
+    public PasswordEncoder passwordEncoder() { return new BCryptPasswordEncoder(); }
+
     @Bean
-    public PasswordEncoder passwordEncoder() {
-        return new BCryptPasswordEncoder();
+    public AuthenticationManager authenticationManager(AuthenticationConfiguration cfg) throws Exception {
+        return cfg.getAuthenticationManager();
     }
 }
