@@ -1,22 +1,11 @@
-// package com.csse.smartwaste.repository;
-
-// import com.csse.smartwaste.model.User;
-// import org.springframework.data.mongodb.repository.MongoRepository;
-// import java.util.Optional;
-
-// public interface UserRepository extends MongoRepository<User, String> {
-//     Optional<User> findByUsername(String username);
-//     boolean existsByUsername(String username);
-// }
 package com.csse.smartwaste.repository;
-
-import com.csse.smartwaste.model.User;
-import org.springframework.data.mongodb.repository.MongoRepository;
-import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
 
-@Repository
+import org.springframework.data.mongodb.repository.MongoRepository;
+
+import com.csse.smartwaste.model.User;
+
 public interface UserRepository extends MongoRepository<User, String> {
     Optional<User> findByUsername(String username);
     boolean existsByUsername(String username);
